Moved query creation methods into queryCreator object
